fixed gradle
