+++ .gitignore
@@ -14,3 +14,4 @@ hs_err_pid*
 # Gradle stuff
 build/
 .gradle/
+/.nb-gradle/
