+++ src/main/java/at/unisalzburg/apted/node/Node.java
@@ -21,7 +21,7 @@
  * SOFTWARE.
  */
 
-package node;
+package at.unisalzburg.apted.node;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -87,7 +87,7 @@ public class Node<D> {
   /**
    * Returns a string representation of the tree in bracket notation.
    *
-   * <p>IMPORTANT: Works only for nodes storing {@link node.StringNodeData}
+   * <p>IMPORTANT: Works only for nodes storing {@link at.unisalzburg.apted.node.StringNodeData}
    * due to using {@link node.StringNodeData#getLabel()}.
    *
    * @return tree in bracket notation.
