+++ src/main/java/at/unisalzburg/apted/node/StringNodeData.java
@@ -21,7 +21,7 @@
  * SOFTWARE.
  */
 
-package node;
+package at.unisalzburg.apted.node;
 
 /**
  * Represents a node label that consists of a single string value. Such label
